+++ nestjs-BE/server/src/boards/dto/create-board.dto.ts
@@ -1,6 +1,6 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { IsNotEmpty, IsString, MaxLength } from 'class-validator';
-import { MAX_NAME_LENGTH } from '../../config/magic-number';
+import { MAX_NAME_LENGTH } from '../../config/constants';
 
 export class CreateBoardDto {
   @ApiProperty({ description: '보드 이름' })
