+++ nestjs-BE/server/src/spaces/dto/create-space.dto.ts
@@ -1,6 +1,6 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { IsNotEmpty, IsString, MaxLength } from 'class-validator';
-import { MAX_NAME_LENGTH } from '../../config/magic-number';
+import { MAX_NAME_LENGTH } from '../../config/constants';
 import { Expose } from 'class-transformer';
 import { v4 as uuid } from 'uuid';
